add junit coverage for the IdempotencyTaskWorker.java
